Rename package from mquickjs_python to microjs
- Rename src/mquickjs_python to src/microjs
- Update pyproject.toml: package name and description
- Update README.md: title and import examples
- Update spec.md: title and file structure
- Update open-problems.md: title references
- Update all test files with new import paths
- All 458 tests pass (443 passed, 15 xfailed)
