+++ src/microjs/__init__.py
@@ -1,10 +1,10 @@
 """
-MQuickJS Python - A Pure Python JavaScript Sandbox Engine
+microjs - A Pure Python JavaScript Sandbox Engine
 
 A sandboxed JavaScript execution environment with memory and time limits,
 implemented entirely in Python with no external dependencies.
 
-Based on: https://github.com/bellard/mquickjs
+Inspired by: https://github.com/bellard/mquickjs
 """
 
 __version__ = "0.1.0"
