+++ src/microjs/lexer.py
