+++ src/microjs/opcodes.py
