+++ src/microjs/regex/opcodes.py
